+++ js/main.js
@@ -1,13 +1,11 @@
 import {createOffers} from './data.js';
 import {COUNT_OF_AD} from './constants.js';
 import {createOfferCard} from './offer-card.js';
-import {
-  disablePage,
-  enablePage
-} from './form-activity-state.js';
+import {changeFormsState} from './toggle-status-page.js';
 
-disablePage();
+changeFormsState(false);
 const similarOffers = createOffers(COUNT_OF_AD);
 
 createOfferCard(similarOffers[0]);
-enablePage();
+
+changeFormsState(true);
